training/requirements.txt
numpy
pillow
requests
torch
torchvision
scikit-learn
sqlalchemy
psycopg2-binary
pyyaml
boto3
pika
pydantic
pandas
matplotlib
